requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.47.0
django-cleanup==5.2.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0
